Update UpdatePatch.py
This file, is meant to be running for version 3.2 and below, but it can be used for any version.
It's just an update manager, since the older versions had offline update system, this one has online, so this is just an extension of the older version's update functionality.
Note that you need to have the version file ready, because this uses that file to check if the program is old or not.